Removed ropevim from my vim install. To reinstall follow steve losh's instructions here: http://forrst.com/posts/install_ropevim_sh-77c.
Per Steve:
Installing Ropevim is a pretty big fustercluck of dependencies and path
issues, so here's a script that does it for you.

I've tweaked the original script to account for Vim's "python3" support and
to use my fork of rope, which fixes a bug that happens when you use it in a
Mercurial repo with Mercurial's progress extension turned on.

I suggest running the following to create a separate place for it in your
~/.vim folder, since it's not a Pathogen bundle:

mkdir -p ~/.vim/sadness/ropevim
cd ~/.vim/sadness/ropevim
~/Downloads/install_ropevim.sh