Always start on first line of git commit message.
When editing a git commit message (.git/COMMIT_EDITMSG) you often won't
start on the first line due to Vim remembering your last position in
that file. An easy way to fix this was to add the following line to my
vimrc.

More information can be found at this link
http://vim.wikia.com/wiki/Always_start_on_first_line_of_git_commit_message