Add stylus yasnippets
The main Stylus plugin used to contain the snippets in this plugin, but was separated out into two, to allow people to use the recommended snippet expander emmet for a better coding experience. This package is maintained for users who for some reason do not want to use another snippet expander, and only want the stylus snippets that used to be in the main package.