set evil-move-beyond-eol to t
This variable allows to put the point behind the last character of a line in normal state. So now eval-last-sexp behaves as expected in elisp and all other languages.

Note that this is not the default in Vim but it does not make a lot of sense to force this behavior in Emacs where eval-last-sexp is so important.

Anyway it is easy to revert the behavior by setting this variable back to nil in dotspacemacs/config