Enabled comment inversion without toggling switch
Previously, you had to set the evilnc-invert-comment-line-by-line to be
enable inverted comments. Since whether you want to invert comments or
not varies on a case to case basis, depending on the region you wish to
comment, this is bad UI.

Made two functions per original commenting function, one for inverse the
other for regular.

Added keybindings for the new inverted operations on capital letter, so
that comment lines is <Leader> c l and invert comment lines is <Leader>
c L.

I removed the comment invert toggle as it is no longer needed.

Also add as the key bindings also import "gc" and "gy" bindings from
evil-commentary