Rewrite emacs-leader handling
This is how evil-leader gets `<Leader> m` to work for mode specific
bindings, and we need to mirror this step for `M-m`. This is big change,
but in order for `M-m` to work properly we need to remap it to the
mode-map whenever possible.

Additional fix for evil-org whose config was unbinding important leader
keys in org-mode