Focus current file when opening project's neotree
The `, p t` keybinding opens the neotree at the project root. Currently the neotree only shows the root of the project, nothing is expanded and no file is focused/selected. With this modification, the neotree is still showing the project root, but with the current file being focused/selected in the hierarchy (its path being expanded).