Give a prefix and bindings for org applications
Use `, a o` as prefix for org related applications like `org-agenda` or `org-capture`. The `a` is for "application", and `o` for "org". To avoid long keystrokes for `agenda`, direct bindings are offered through
`, a o <key>`.

Because `org-agenda` is one of the principal application, it has be mapped to repeating the last letter (`, a o o`).