Adapt project-search keybindings
For buffers, `*` search for the word under the cursor, and `/` ask to search for an expression. To keep the same mnemonic at project level, this commit change `SPC /` to ask for an expression to search, and add
`SPC *` to search for the expression under the cursor (in fact the only difference being the expression pre-filled inside the input field).