feat: unprettify symbol at point
In Emacs 25.1, thanks to Tassilo Horn, there’s a new variable called
`prettify-symbols-unprettify-at-point`. If you set it to `t`,
`prettify-symbols` will “unprettify” a symbol as long as the cursor is
inside it. So you can easily (and temporarily) reveal a symbol by just
moving over to it.

My preference is to set it to right-edge which also reveals the symbol
if the point is immediately after it.